git-svn-id: https://source.openmpt.org/svn/openmpt/branches/1.26-maintenance@6990 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/View_tre.h
@@ -146,8 +146,8 @@ protected:
 		DlsItem(uint32 bank, uint32 itemData) : ModItem(MODITEM_DLSBANK_INSTRUMENT, itemData | ((bank << (DLS_HIBANKSHIFT - 16)) & DLS_HIBANKMASK), (uint16)bank) { }
 
 		uint32 GetBankIndex() const { return val2 | ((val1 & DLS_HIBANKMASK) >> (DLS_HIBANKSHIFT - 16)); }
-		uint8 GetRegion() const { return static_cast<uint8>((val1 & DLS_REGIONMASK) >> DLS_REGIONSHIFT); }
-		uint8 GetInstr() const { return static_cast<uint8>((val1 & DLS_INSTRMASK)); }
+		uint32 GetRegion() const { return (val1 & DLS_REGIONMASK) >> DLS_REGIONSHIFT; }
+		uint32 GetInstr() const { return (val1 & DLS_INSTRMASK); }
 		bool IsPercussion() const { return ((val1 & DLS_TYPEMASK) == DLS_TYPEPERC); }
 		bool IsInstr() const { return ((val1 & DLS_TYPEMASK) == DLS_TYPEINST); }
